java project with no built in build but can run
login in work but enquiry and project might have problems with the loading part, need to check but can login in for now for the rest havent debugged if the options for the roles work. Main menu for now dont have choice to login in as which role for now.